Victory Condition Functionality
Added the ability to keep track of when a character's HP hits zero.

The attack method will now properly subtract the amount of damage taken from the characters health.

During each turn the game will now actively check when either side is completely defeated before announcing the winner.

Included the option to allow the player to choose to play again.